refactor: 스토리북 경로 수정
