+++ src/stories/RegisterEmailInput.stories.tsx
@@ -1,6 +1,6 @@
 import { Meta, StoryObj } from '@storybook/react';
 
-import EmailInput from '~/pages/RegisterPage/components/EmailInput';
+import EmailInput from '~/pages/register/components/EmailInput';
 
 export default {
   title: 'Common/Components/EmailInput',
